inference/requirements.txt
# Core dependencies (always needed)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# vLLM (only install on Linux with GPU)
# vllm>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
